requirements.txt
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
notebook==7.0.6
plotly==5.18.0
scikit-learn==1.3.2
statsmodels==0.14.1
python-dotenv==1.0.0
pytest==7.4.3